Adapted tests to new merge
